first refactoring for aggregate update
