+++ blimp/utils.py
@@ -1,11 +1,12 @@
 """Miscellaneous utilities for blimp"""
-from typing import List, Union, Optional
+from typing import List, Tuple, Union, Optional
 import logging
 
 from welford import Welford
 from aicsimageio import AICSImage
 import numpy as np
 import dask.array as da
+import skimage.measure
 
 logger = logging.getLogger(__name__)
 
@@ -86,6 +87,34 @@ def convert_array_dtype(arr, dtype, round_floats_if_necessary=False, copy=True):
                     + f" minimum data type necessary to hold the values is ({new.dtype}) "
                 )
     return new
+
+
+def convert_image_dtype(image: AICSImage, dtype: np.dtype) -> AICSImage:
+    """
+    Change the dtype of an AICSImage object.
+
+    This function converts the data type of the image's data using the
+    `convert_array_dtype` function. It preserves the channel
+    names and physical pixel sizes of the original image.
+
+    Parameters
+    ----------
+    image : AICSImage
+        The input AICSImage object containing the image data.
+    dtype : np.dtype
+        The target dtype to convert the image data to.
+
+    Returns
+    -------
+    AICSImage
+        A new AICSImage object with the updated dtype,
+        retaining the original channel names and physical pixel sizes.
+    """
+    return AICSImage(
+        convert_array_dtype(image.data, dtype),
+        channel_names=image.channel_names,
+        physical_pixel_sizes=image.physical_pixel_sizes,
+    )
 
 
 def equal_dims(a, b, dimensions="TCZYX"):
@@ -223,7 +252,9 @@ def check_correct_dimension_order(images: Union[AICSImage, List[AICSImage]]) -> 
 
 
 def check_uniform_dimension_sizes(
-    images: Union[AICSImage, List[AICSImage]], omit: Union[int, str, None] = None, check_dtype: bool = True,
+    images: Union[AICSImage, List[AICSImage]],
+    omit: Union[int, str, None] = None,
+    check_dtype: bool = True,
 ) -> bool:
     """
     Check that the dimensions of a list of AICSImages are uniform.
@@ -296,35 +327,6 @@ def check_uniform_dimension_sizes(
     return result
 
 
-def change_image_dtype(
-    image: AICSImage,
-    dtype: np.dtype
-) -> AICSImage:
-    """
-    Change the dtype of an AICSImage object.
-
-    This function converts the data type of the image's data using the 
-    `convert_array_dtype` function. It preserves the channel 
-    names and physical pixel sizes of the original image.
-
-    Parameters
-    ----------
-    image : AICSImage
-        The input AICSImage object containing the image data.
-    dtype : np.dtype
-        The target dtype to convert the image data to.
-
-    Returns
-    -------
-    AICSImage
-        A new AICSImage object with the updated dtype, 
-        retaining the original channel names and physical pixel sizes.
-    """
-    return AICSImage(convert_array_dtype(image.data, dtype),
-                     channel_names=image.channel_names,
-                     physical_pixel_sizes=image.physical_pixel_sizes)
-
-
 def mean_std_welford(images: List[AICSImage], log_transform: bool = False) -> tuple:
     n_c = images[0].dims.C
     w = [Welford() for i in range(n_c)]
@@ -336,12 +338,12 @@ def mean_std_welford(images: List[AICSImage], log_transform: bool = False) -> tu
             if log_transform:
                 is_zero = array == 0
                 if np.any(is_zero):
-                    logger.warn("image contains zero values")
+                    logger.warning("image contains zero values")
                 array = np.log10(array)
                 # The log10 transform sets zero pixel values to -inf
                 array[is_zero] = 0
             if np.any(np.isinf(array)):
-                logger.warn("skip image because it contains infinite values")
+                logger.warning("skip image because it contains infinite values")
             else:
                 w[c].add(array)
 
@@ -511,7 +513,7 @@ def concatenate_images(
         raise TypeError("Cannot concatenate list elements of different types")
 
     if isinstance(images, AICSImage):
-        logger.warn("Tried to concatenate a single image, result is unchanged")
+        logger.warning("Tried to concatenate a single image, result is unchanged")
         return images
 
     elif not dimension_sizes_match:
@@ -519,7 +521,7 @@ def concatenate_images(
 
     channel_names = images[0].channel_names
     physical_pixel_sizes = images[0].physical_pixel_sizes
-    dtype = images[0].dtype
+    images[0].dtype
 
     if order == "append":
         logger.debug(f"Concatenating by appending, on axis {axis}")
@@ -696,3 +698,328 @@ def estimate_focus_plane(
             max_pos = np.argmax(f4)
 
     return int(max_pos)
+
+
+def get_channel_names(image: AICSImage, input: Optional[Union[int, str, List[Union[int, str]]]] = None) -> List[str]:
+    """
+    Get the channel names based on input.
+
+    Parameters
+    ----------
+    image
+        An instance of AICSImage with channel_names attribute.
+    input
+        The input specifying channels as integer(s) or string(s).
+        Defaults to None.
+
+    Returns
+    -------
+    A list of strings representing the requested channel names, without duplicates,
+    or the full list of channel names if input is None.
+
+    Raises
+    ------
+        ValueError: If input is not of appropriate type or contains invalid channel names.
+    """
+
+    if input is None:
+        return image.channel_names
+
+    if not isinstance(input, (int, str, list)):
+        raise ValueError("Input must be an integer, a string, a list of integers/strings, or None.")
+
+    if isinstance(input, int):
+        if input not in range(len(image.channel_names)):
+            raise ValueError(f"Integer input for channel name must be in the range 0 to {len(image.channel_names)}.")
+        return [image.channel_names[input]]
+
+    elif isinstance(input, str):
+        if input not in image.channel_names:
+            raise ValueError("String input for channel name must be a valid channel name.")
+        return [input]
+
+    elif isinstance(input, list):
+        channel_names = []
+        for item in input:
+            if isinstance(item, int):
+                if item not in range(len(image.channel_names)):
+                    raise ValueError(
+                        f"Integer input for channel name must be in the range 0 to {len(image.channel_names)}."
+                    )
+                channel_names.append(image.channel_names[item])
+            elif isinstance(item, str):
+                if item not in image.channel_names:
+                    raise ValueError("String input must be a valid channel name.")
+                channel_names.append(item)
+            else:
+                raise ValueError("List elements must be integers or strings.")
+        return list(set(channel_names))  # Remove duplicates
+
+
+def cropped_array_containing_object(array: np.ndarray, bboxes: list, label: int, pad: int = 1) -> np.ndarray:
+    """
+    Extract a region from the input array corresponding to a single object.
+
+    Parameters
+    ----------
+    array
+        The input array from which the object intensity will be extracted.
+    bboxes
+        List containing [start_row, end_row, start_column, end_column] of the bounding box.
+    label
+        The label of the object for which the intensity needs to be extracted.
+
+    Returns
+    -------
+    numpy.ndarray
+        Intensity values of the specified object extracted from the intensity image.
+
+    Raises
+    ------
+    ValueError
+        If the array is None.
+
+    Notes
+    -----
+    The function uses the provided bounding box coordinates to extract the region of
+    interest (ROI) corresponding to the specified object label. It then optionally pads
+    the extracted ROI and returns it as a separate image.
+
+    Examples
+    --------
+    >>> array = np.array([[0, 1, 2], [3, 4, 5], [6, 7, 8]])
+    >>> bboxes = {'object1': [0, 2, 0, 2]}
+    >>> label = 'object1'
+    >>> object_intensity = get_object_array(array, bboxes, label)
+    >>> print(object_intensity)
+    [[0 1 2]
+     [3 4 5]
+     [6 7 8]]
+    """
+    if array is None:
+        raise ValueError("No intensity image available.")
+    bbox = bboxes[label]
+    return extract_bbox(array, bbox=bbox, pad=pad)
+
+
+def extract_bbox(array: np.ndarray, bbox: list, pad: int = 0) -> np.ndarray:
+    """
+    Extracts a bounding box region from the given image.
+
+    Parameters
+    ----------
+    array : numpy.ndarray
+        The input image from which the bounding box region will be extracted.
+    bbox : list
+        List containing [start_row, end_row, start_column, end_column] of the bounding box.
+    pad : int, optional
+        The number of pixels to pad around the extracted region, by default 0.
+
+    Returns
+    -------
+    numpy.ndarray
+        The extracted bounding box region from the input image.
+
+    Notes
+    -----
+    This function extracts the region defined by the provided bounding box coordinates
+    from the input image. Optionally, padding can be applied around the extracted region.
+
+    Examples
+    --------
+    >>> image = np.array([[1, 2, 3], [4, 5, 6], [7, 8, 9]])
+    >>> bbox = [0, 1, 0, 2]
+    >>> padded_bbox = extract_bbox(image, bbox, pad=1)
+    >>> print(padded_bbox)
+    [[0 0 0 0 0]
+     [0 1 2 3 0]
+     [0 4 5 6 0]
+     [0 0 0 0 0]]
+    """
+    cropped_array = array[bbox[0] : bbox[1], bbox[2] : bbox[3]]
+    if pad:
+        cropped_array = np.lib.pad(cropped_array, (pad, pad), "constant", constant_values=(0))
+    return cropped_array
+
+
+def _object_mips_3D(
+    intensity_array: np.ndarray,
+    label_array: np.ndarray,
+) -> List:
+    """
+    Generate Maximum Intensity Projections (MIPs) for labeled 3D regions.
+
+    This function computes the MIP for each labeled region in a 3D intensity array.
+    It identifies individual regions based on the provided label array, crops the
+    corresponding regions from the intensity array, masks out regions not belonging
+    to the current label, and then calculates the MIP along the z-axis for each region.
+
+    Parameters
+    ----------
+    intensity_array : np.ndarray
+        A 3D numpy array representing the intensity values of the volume.
+    label_array : np.ndarray
+        A 3D numpy array with the same shape as `intensity_array`, where each unique
+        integer label identifies a different region in the volume.
+
+    Returns
+    -------
+    List
+        A list of 2D numpy arrays, each representing the MIP of a labeled region along
+        the z-axis.
+    """
+
+    regions = skimage.measure.regionprops(label_array)
+    mips = [None] * len(regions)
+
+    for index, region in enumerate(regions):
+        bbox = region.bbox
+        object_crop = intensity_array[bbox[0] : bbox[3], bbox[1] : bbox[4], bbox[2] : bbox[5]]
+        mask_crop = label_array[bbox[0] : bbox[3], bbox[1] : bbox[4], bbox[2] : bbox[5]]
+        object_crop_masked = np.copy(object_crop)
+        object_crop_masked[mask_crop != index + 1] = 0
+        mips[index] = object_crop_masked.max(axis=0)
+    return mips
+
+
+def _object_middle_planes_3D(intensity_array: np.ndarray, label_array: np.ndarray) -> List:
+    """
+    Extract middle planes of labeled 3D regions.
+
+    This function computes the middle plane for each labeled region in a 3D intensity array.
+    It identifies individual regions based on the provided label array, determines the middle
+    plane along the z-axis for each region, crops the corresponding region from the intensity
+    array, masks out regions not belonging to the current label, and then stores the masked
+    middle plane.
+
+    Parameters
+    ----------
+    intensity_array : np.ndarray
+        A 3D numpy array representing the intensity values of the volume.
+    label_array : np.ndarray
+        A 3D numpy array with the same shape as `intensity_array`, where each unique
+        integer label identifies a different region in the volume.
+
+    Returns
+    -------
+    List
+        A list of 2D numpy arrays, each representing the masked middle plane of a labeled
+        region along the z-axis.
+    """
+    regions = skimage.measure.regionprops(label_array)
+    meds: List[Optional[np.ndarray]] = [None] * len(regions)
+    for index, region in enumerate(regions):
+        bbox = region.bbox
+        object_crop = intensity_array[int(np.rint((bbox[0] + bbox[3]) / 2.0)), bbox[1] : bbox[4], bbox[2] : bbox[5]]
+        mask_crop = label_array[int(np.rint((bbox[0] + bbox[3]) / 2.0)), bbox[1] : bbox[4], bbox[2] : bbox[5]]
+        object_crop_masked = np.copy(object_crop)
+        object_crop_masked[mask_crop != index + 1] = 0
+        meds[index] = object_crop_masked
+    return meds
+
+
+def concatenated_projection_image_3D(
+    intensity_image: AICSImage,
+    label_image: AICSImage,
+    label_channel: int = 0,
+    label_name: str = "Nuclei-3D-MIP",
+    projection_type: str = "MIP",
+) -> Tuple[AICSImage, AICSImage]:
+    """
+    Generate concatenated projections (MIPs or middle planes) for 3D intensity and label images.
+
+    This function computes either the Maximum Intensity Projections (MIPs) or middle planes
+    for the labeled regions in a 3D intensity image, and concatenates them to form a single
+    2D image per channel. It also creates a concatenated projection image for the label image itself.
+
+    This can be used to compute intensity feature values using the standard quantification pipeline
+    for 2D images.
+
+    Parameters
+    ----------
+    intensity_image : AICSImage
+        AICSImage object containing the 3D intensity data.
+    label_image : AICSImage
+        AICSImage object containing the 3D label data.
+    label_channel : int, optional
+        The channel index in `label_image` to be used for generating the label projection,
+        by default 0.
+    projection_type : str, optional
+        Type of projection to generate. Either "MIP" for Maximum Intensity Projection
+        or "middle" for middle plane projection, by default "MIP".
+
+    Returns
+    -------
+    Tuple[AICSImage, AICSImage]
+        A tuple containing two AICSImage objects: the concatenated projection of the intensity
+        image and the concatenated projection of the label image.
+
+    Raises
+    ------
+    ValueError
+        If the dimensions of `label_image` and `intensity_image` do not match.
+        If either `label_image` or `intensity_image` has Z-dimension size <= 1.
+        If `projection_type` is not "mip" or "middle".
+    """
+
+    # Check if dimensions of intensity_image and label_image match
+    if label_image.dims[["T", "Z", "Y", "X"]] != intensity_image.dims[["T", "Z", "Y", "X"]]:
+        raise ValueError("The dimensions of label_image and intensity_image do not match.")
+
+    # Check if the Z-dimension is greater than 1
+    if label_image.dims.Z <= 1:
+        raise ValueError("Both label_image and intensity_image must have a Z-dimension greater than 1.")
+
+    # Determine the projection function and label name based on projection_type
+    if projection_type == "MIP":
+        projection_func = _object_mips_3D
+        label_name = label_name
+    elif projection_type == "middle":
+        projection_func = _object_middle_planes_3D
+        label_name = label_name
+    else:
+        raise ValueError("projection_type must be either 'mip' or 'middle'.")
+
+    # Generate label projection arrays
+    label_projection_arrays = projection_func(
+        intensity_array=label_image.get_image_data("ZYX", C=label_channel),
+        label_array=label_image.get_image_data("ZYX", C=label_channel),
+    )
+
+    y_extent = max(m.shape[0] for m in label_projection_arrays)
+
+    # Concatenate label projection arrays
+    concat_label_projection_arrays = np.concatenate(
+        [np.pad(m, pad_width=(y_extent - m.shape[0], 0), constant_values=0) for m in label_projection_arrays], axis=1
+    )
+
+    projection_label_image = AICSImage(
+        concat_label_projection_arrays[np.newaxis, np.newaxis, np.newaxis, :, :],
+        channel_names=[label_name],
+        physical_pixel_sizes=label_image.physical_pixel_sizes,
+    )
+
+    # Generate intensity projection arrays and concatenate them
+    concat_intensity_projection_arrays = [
+        np.concatenate(
+            [
+                np.pad(m, pad_width=(y_extent - m.shape[0], 0), constant_values=0)
+                for m in projection_func(
+                    intensity_array=intensity_image.get_image_data("ZYX", C=channel),
+                    label_array=label_image.get_image_data("ZYX", C=label_channel),
+                )
+            ],
+            axis=1,
+        )
+        for channel in range(intensity_image.dims.C)
+    ]
+
+    concat_intensity_projection_stack = np.stack(concat_intensity_projection_arrays, axis=0)
+
+    projection_intensity_image = AICSImage(
+        concat_intensity_projection_stack[np.newaxis, :, np.newaxis, :, :],
+        channel_names=intensity_image.channel_names,
+        physical_pixel_sizes=intensity_image.physical_pixel_sizes,
+    )
+
+    return projection_intensity_image, projection_label_image
